requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
pytesseract==0.3.10
Pillow==10.4.0
pdfplumber==0.11.4
python-docx==1.1.2
groq==0.9.0
httpx==0.27.0
gunicorn==22.0.0
